refactor(web): merge landing scroll utils and fix gsap integration
