+++ apps/web/components/Feature/landing/SecurePayment/index.tsx
@@ -19,12 +19,12 @@ import { MonoText } from "@/components/UI/Monotext";
 import COLORS from "@repo/ui/colors";
 import ScrollReveal from "@/components/UI/ScrollReveal";
 import ImageReveal from "@/components/UI/ImageReveal";
-import { LANDING_REVEAL } from "@/utils/landingReveal";
+import { LANDING_REVEAL } from "@/utils/landingUtils";
 import { IMAGE_SIZES } from "@/utils/imageSizes";
 import {
   LANDING_IMAGE_FLAGS,
   LANDING_REVEAL_VARIANTS,
-} from "@/utils/landingConfig";
+} from "@/utils/landingUtils";
 
 export default function SecurePaymentSection() {
   const { t } = useTranslation();
