+++ apps/web/package.json
@@ -9,7 +9,6 @@
     "lint": "eslint"
   },
   "dependencies": {
-    "@gsap/react": "^2.1.2",
     "@hookform/resolvers": "^5.2.2",
     "@repo/ui": "workspace:*",
     "@tanstack/react-query": "^5.100.5",
