--- apps/web/utils/landingReveal.ts
@@ -1,14 +1,0 @@
-export const LANDING_REVEAL = {
-  sequenceStepDelay: 0.1,
-  ctaSubtitleDelay: 0.1,
-  ctaSubtitleStepDelay: 0.1,
-  ctaCardStaggerDelay: 0.08,
-  shortDelay: 0.1,
-  mediumDelay: 0.2,
-  revealDuration: 1.2,
-  longRevealDuration: 1.4,
-  extraLongRevealDuration: 1.6,
-  heroKenBurnsDuration: 1.8,
-  heroStart: "top 100%",
-  imageRevealStart: "top 88%",
-} as const;
