--- apps/web/utils/safeImage.ts
@@ -1,8 +1,0 @@
-export const SAFE_IMAGE_FALLBACKS = {
-  default: "/images/fallbacks/default.svg",
-  avatar: "/images/fallbacks/avatar.svg",
-  thumbnail: "/images/fallbacks/thumbnail.svg",
-  hero: "/images/fallbacks/hero.svg",
-} as const;
-
-export const SAFE_IMAGE_DECODING = "async" as const;
